+++ UnitTests/BraintreePayPalTests/BTTokenizationService_PayPal_Tests.swift
@@ -1,9 +1,0 @@
-import XCTest
-
-class BTTokenizationService_PayPal_Tests: XCTestCase {
-    func testSingleton_hasPayPalTypeAvailable() {
-        let sharedService = BTTokenizationService.shared()
-
-        XCTAssertTrue(sharedService.isTypeAvailable("PayPal"))
-    }
-}